Clean up TryFromTsStd trait
